requirements.txt
streamlit==1.11.0
torch==1.12.0
torchvision==0.13.0
Pillow==9.2.0
altair==4.1.0
torch==1.12.0
torchvision==0.13.0
Pillow==9.2.0